+++ core/src/main/java/org/patternfly/popper/TriggerAction.java
@@ -23,5 +23,7 @@ public enum TriggerAction {
 
     click,
 
+    stayOpen,
+
     manual
 }