+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Add
+
+- MultiSelect component (part of #68)
+
 ## [0.2.9] - 2024-08-21
 
 ### Add